Async Wrapper
Refactored the code to remove the multiple try-catch block on all the controller method. Using the controller method as a parameter, the asyncWrapper middleware will handle the try-catch and then if there is an error, the next function will handle it. The next function to implement now is the next function middleware.